Refactor Redis connection pool to permit sharing.
