+++ nodes/core/lib/redisConnectionPool.js
@@ -1,0 +1,52 @@
+/**
+ * Copyright 2013, 2014 IBM Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+
+module.exports = (function() {
+    "use strict";
+    var redis = require("redis");
+    var util = require("util");
+
+    var connections = {};
+    var obj = {
+        get: function(host,port) {
+            var id = host+":"+port;
+            if (!connections[id]) {
+                connections[id] = redis.createClient(port,host);
+                connections[id].on("error",function(err) {
+                    util.log("[redis] "+err);
+                });
+                connections[id].on("connect",function() {
+                    util.log("[redis] connected to "+host+":"+port);
+                });
+                connections[id]._id = id;
+                connections[id]._nodeCount = 0;
+            }
+            connections[id]._nodeCount += 1;
+            return connections[id];
+        },
+        close: function(connection) {
+            connection._nodeCount -= 1;
+            if (connection._nodeCount === 0) {
+                if (connection) {
+                    clearTimeout(connection.retry_timer);
+                    connection.end();
+                }
+                delete connections[connection._id];
+            }
+        }
+    };
+    return obj;
+})();
